python/requirements.txt
numpy
scipy==1.4.1
pyaudio
PyQt5
pyqtgraph
pyserial